+++ src/bare-metal/sf2/rodos/hal/hal_gpio.cpp
@@ -44,7 +44,7 @@ HAL_GPIO::HAL_GPIO(GPIO_PIN pinIdx) {
 }
 
 
-int32_t HAL_GPIO::init(bool isOutput, uint32_t numOfPins, uint32_t initVal) {
+int32_t HAL_GPIO::init(bool isOutput, uint8_t numOfPins, uint32_t initVal) {
     if(numOfPins > 0) {
         context->numOfPins = numOfPins; // numOfPins has to be > 0 -> if new value is 0 keep the default value
     }
@@ -77,7 +77,7 @@ int32_t HAL_GPIO::config(GPIO_CFG_TYPE type, uint32_t paramVal) {
 
         case GPIO_CFG_NUM_OF_PINS:
             if((paramVal > 0) && (paramVal < 32)) { // numOfPins has to be > 0 and < 256 -> uint8_t
-                context->numOfPins = paramVal;
+                context->numOfPins = static_cast<uint8_t>(paramVal);
                 context->setPinMask();
                 return 0;
             }
