+++ src/bare-metal/sf2/rodos/hw_tcp.h
@@ -1,7 +1,7 @@
 #pragma once
 
 #include "lwip/tcp.h"
-#include <stddef.h>
+#include <stdint.h>
 
 err_t tcp_recv_func(void* arg, struct tcp_pcb* tpcb, struct pbuf* p, err_t);
 err_t tcp_accept_func(void* arg, struct tcp_pcb* newpcb, err_t err);
@@ -13,8 +13,8 @@ class TCPBase {
   public:
     virtual ~TCPBase() { tcp_abort(pcb); }
 
-    virtual int sendData(void* buf, size_t len);
-    virtual int getData(void* buf, size_t maxLen);
+    virtual int sendData(void* buf, uint16_t len);
+    virtual int getData(void* buf, uint16_t maxLen);
 
   protected:
     pbuf*    recvBuf   = nullptr;
@@ -22,7 +22,7 @@ class TCPBase {
     int      errorCode = -1;
 
   protected:
-    size_t readPos = 0;
+    uint16_t readPos = 0;
 
     friend err_t ::tcp_accept_func(void* arg, struct tcp_pcb* newpcb, err_t err);
     friend err_t ::tcp_recv_func(void* arg, struct tcp_pcb* tpcb, struct pbuf* p, err_t);
@@ -33,10 +33,10 @@ class HW_TCPServer : public TCPBase {
   public:
     ~HW_TCPServer();
 
-    bool listen(const long portNr);
+    bool listen(const int32_t portNr);
     bool acceptNewConnection();
-    int  sendData(void* buf, size_t len) override;
-    int  getData(void* buf, size_t maxLen) override;
+    int  sendData(void* buf, uint16_t len) override;
+    int  getData(void* buf, uint16_t maxLen) override;
     int  getErrorCode() const;
 
     tcp_pcb* listenPcb = nullptr;
@@ -46,9 +46,9 @@ class HW_TCPClient : public TCPBase {
   public:
     ~HW_TCPClient();
 
-    bool reopen(const long portNr, const char* hostname = "localhost");
-    int  sendData(void* buf, size_t len) override;
-    int  getData(void* buf, size_t maxLen) override;
+    bool reopen(const int32_t portNr, const char* hostname = "localhost");
+    int  sendData(void* buf, uint16_t len) override;
+    int  getData(void* buf, uint16_t maxLen) override;
     int  getErrorCode() const;
 };
